Fix S32-hash/delete-adverb.t and squish.t if hash keys iter randomized
Soon I will be making MoarVM randomize the order of the iterator. Though
we have a random order each start of MoarVM currently due to the random
hash secret, if you make two hashes and add things in the same order you
will get the same order as each other when you do .keys. This will be
changing soon and you can't rely on the order (unless you don't modify
the hash, in that case order is consistent).