Correct tests using return instead of return-rw
Even if the enclosing sub is marked `is rw` or `is raw`, it is still
required to `return-rw` the `Proxy`. We got away with this so far, as
Rakudo had an implementation limitation/bug that meant a `Proxy` being
returned in a non-rw context would not be stripped from around the
value. A Rakudo branch currently in progress fixes the issue. The
corrected test works in both master and in the branch. However, the
branch also fixes a couple of todo'd tests in one of these files too,
which can then be un-todo'd once the branch merges.