Remove over-specific and dubious error test
For one, this test checked for X::AdHoc, which a test should never do.
It also checked that the error message contained the name of a readonly
being assigned to (in parens too, for that matter). While we could in
this particular case do that in Rakudo master, in numerous cases the
information was not around at runtime to do this. Changes to Scalar
container representation mean that we simply don't carry that around
ever now.

This kind of error should really be detected at compile time and
reported then. That will allow us to do it and include variable name
and location for many more cases. However, it will need to be done as
part of a language version bump, since it can break compilation of
existing code that just never touched the broken code path. At that
point, tests can be added to cover the compile-time reporting, which
should use a typed exception to allow better testing.