[v6.d REVIEW] Fix bad "dir | say" test
The test specs behaviour that IMo should not be specced at all
and that is that `say` without parens throws a P5ism error.

However, reading the OP on the ticket this test is meant to cover,
we can clearly see the original crash was due to slipping of a Bool
into argument lists and nothing to do with neither `say` nor `dir`.

Orig: https://github.com/perl6/roast/commit/1c1d01eaf