Revert "Add  test for loading module .pm files (#466)"
This reverts commit 5c7df0d522a24d352ae229ea75f277fa8310b7a1.

As mentioned in my previous commit I do not think this is appropriate
to test in the roast. i.e. the tests we replaced were not for
explicitly testing what file extension works, so its not a single
persons place to decide that now the spec incorporates specific
file extensions in this instance.