Change expectation of non-normalization of ZDRs
BDFL's blessing: http://colabti.org/irclogger/irclogger_log/perl6-dev?date=2018-09-21#l248

What follows is a rationalization for the change, although it can be equally
construed as mental gymnastics to justify a direct specification violation.

The test in question specs a tiny portion of the behaviour of
ZDRs (Zero Denominator Rationals), and the spec does not concern itself with
any of the other behaviours. Some of those turned out to be buggy and were
fixed. Others, are still buggy and need fixing. While a certain fix is
possible while maintaing this test in passing state, that fix inflicts a
non-trivial performance penalty on many operations on *all* Rationals.

Therefore, I argue the 6.c spec is wrong to spec the behaviour of ZDRs in
such a sparse manner, without any regard to how an implementation can
meet the goals of this particular test, while maintaining bug-free behaviour
of ZDRs. Ergo, the test is wrong and shall be removed.