[v6.d REVIEW] Appendicize questionable `int` limits test
It might be fine, but I'm unsure whether it's accurate to say
`int` can be 32 bits on 32-bit platforms (and it isn't on MoarVM),
which the text of the test suggests.

Shove it into appendices, until our natives are more solidly implemented.

Orig: https://github.com/perl6/roast/commit/f0dbe1c97