Update Uni Normalization to v11.0. Make Uni test doc uniform and clear
Update the normalization tests to Unicode v11.0.

Make the document which tracks the different scripts and tests which
need to be regenerated more uniform with a clear structure.

Add the exact commands to run as well as showing which test files are
generated from each script. Only one or two left need this info added
now.