Add test for decont in ENTER phaser
In a previous version of rakudo, when a block had no local variables,
the ENTER phaser would not successfully decont a variable from a
surrounding scope. This didn't have a ticket and was apparently fixed
in rakudo as a side-effect of 541a4f1628.

Cf. https://colabti.org/irclogger/irclogger_log/perl6?date=2019-04-26#l196