Fix test for JSON exception output of module load failure
The test was overly specific as it tested the serialized JSON string for a
specific class name which fails as soon as the exception gets some role mixed
in.
The test was also underly specific in that it didn't catch a deprecation
message attached to the JSON output.
Fix by using a real JSON parser.