Remove obsolete test file
This hasn't been meaningfully touched in *12* years.  It tests for
basically a port of Perl's opendir and friends, which do not exist
in Raku (although the P5opendir module *does* try to mimic that).
In any case, this test-file is not being run anyway, and such is
only a distraction.