Adapt the test to the syntax and semantics of Raku v6.c
Requires Raku/nqp#650 and rakudo/rakudo#3834 to pass.

I have removed trait `described` because the final implementation of
traits doesn't need covering two different trait routine signatures.

Renamed the role from `description` to `Description` because trait `is`
can't tell the difference between the trait and the role to pun because
of the same name used for both.

Because a trait is applied to a parametric role, not the role group,
test against the first group candidate.