Make `eval` verb a TODO
If somesing is eval-fudged because of a non-implemented or a broken
feature then how do we know when the problem is fixed? I mean a scenario
when the fix done as a result of a ticket being fixed or a similar case.

Now fudge prepends the eval'ed code with a `todo` and instead of `skip`
produces as many `flunk` as many tests were wrapped. So, harness will
inform us about passing TODOs when this happens.