Add testing of mixins under v6.e
Primarly to make sure that cases like `0 but True` are not broken. So,
`mixin-6e.t` is basically a copy of `mixin-6c.t` with few cases removed
where 6.e-declared roles are mixed into core classes. These tests would
require the roles to be declared externally, in a 6.c versioned module
which would eventually wind down to the same testing we already do in
mixin-6c.t.

- Added a test for in-role declared constructor to be invoked upon mixin
  under 6.e