GraphemeBreakTest: fudge lines and norms again text not line numbers
Previously we would fudge specific line numbers for expected failures.
The problem with this is that when the unicode version changes, the line
numbers will often change.

Instead we now fudge the exact lines instead. This should prevent
problems in the future.

Also uncomment the commented out lines in GraphemeBreakTest.txt which
were commented out errronerously. While the test is todo'd, and so will
return 0, it *does* print out the failed lines at the bottom for
informational purposes (and since we really should try and fix the
existing failures... though that is easier said than done). The return
code is what matters.