Make test conform to the new coercion semantics
In support of rakudo/rakudo#4096. Previously the invocation handler was
unconditionally using a coercion method named after the target type. Now
it uses the new coercion protocol. With this change `Str($*USER)` ends
up as `IntStr` because no actual coercion is needed. Correspondingly,
testing for the outcome of it to be a `Str` class makes more sense than
to test for it to be exactly a `Str` instance.