Make sure we have uncached Seqs when checking for exhaustion
With the CachedIterator role, a Seq can be cached even *without* a
specific call to .cache.  These tests where using a List as the
source of the tested Seqs, but with the CachedIterator role these
would be cached be default.  Change the Seqs to uncached by calling
the .slice method on them, which would effectively create an uncached
Seq from the cached Seq.