Fix yet another race condition in kill.t
Most of the time the 'Output correct for Supply.merge on signals' test was not
executed at all because the test program finished normally before it ever got
the signal. Even the test plan did not expect that test to run. Fix by having
the test program basically wait forever for that signal to arrive.