Fix race condition in bug-coverage-stress.t signal test
The signal handler sometimes did not even get a chance to run as the test
program finished right after signalling that it started. To make it reliable
the only proper way to exit the test program must be from the signal handler.
So give that handler a really long time to fire and exit unsuccessfully if it
doesn't.