Adjust troublesome Instant / Duration test
The test depended on the Instant / Durations actually having a Rat
at their core.  Since this may not be the case, an Instant or a Duration
of 2/7 seconds, will not test correctly in real life, as 2/7 seconds
can not actually be represented on any actual hardware now or in the
future (other than by having Rats implemented in hardware I guess).