Add a test of hyper/race iterator method thread-safety
Try running a number of sumultaneous workers trying to obtain a iterator
from the same sequence object. Only one is expected to succeed, the rest
must see `X::Seq::Consumed` thrown.

Since a race condition to happen may require a number of factors to
coincide in time, there is no reliable way to prove it is not there.
But the more repetitions we try the less chances a race is still lurking
somewhere if all attempts succeed. But trying too many times may take
too long on slower systems. As a compromise the test adjusts the number
of repetitions in a way make each test run for approx. 1sec.