Fix a GLR straggler in S01-perl-5-integration/roundtrip.t
Inline::Perl5 has only recently gotten adapted to the flattening behaviour
expected after the GLR. That's why this test's wrong expectations went by
undiscovered for just as long. The test was written years before the GLR
in a time where we flattened arguments as agressively as Perl does. Nowadays
one often needs to be more explicit when expecting that like in this case a
list should be flattened out into a function call's arguments.

Fixes rakudo issue #4418