Modify tests to account for method cache changes
Pre-calculated method caches are not a part of new-disp. The built-in
meta-objects will stop producing them soon on MoarVM; hopefully, with
time, we'll adopt the new-disp approach on JVM too, and can fully
deprecate the method caching related functionality in the metamodel
primitives class.

For now, at least, make the tests not depend on the semantics of the
legacy method caching strategy. Method caches were always just an
optimization, and it was a pragmatic choice to allow user-defined
meta-objects to opt in to it. However, it should not have been
relied upon for semantic effect, which is what these tests did.