Make sure role group delegates .WHY to its default candidate
This required to remove tests which were testing for `Role.WHY`
returning `Nil`. But as was suggested in this comment:

https://github.com/rakudo/rakudo/pull/3549#issuecomment-687587580

the tests were needed to ensure that a documentation comment is attached
to a particular parametric role, not to its group. But since
rakudo/rakudo#3549 makes `.WHY` delegate to the default candidate, these
tests doesn't make sense anymore. Instead why-both.t now makes sure what
we get from a role group is the same object we get from the default
candidate.

Alongside with the essential changes, also unified all tests with regard
to use of idential test assertion subs. They all now are `subtest`
based.