Implement randomized order of indirect smartmatch tests
With new-disp implementation of smartmatch in Rakudo it is now possible
that an outcome of match would be cached and then agressively used when
it shouldn't be. Sometimes such situation my pass unnoticed because of
the order of tests. By randomizing it we reduce the probability of such
situaiton.

Since randomization makes it harder to diagnose and debug a problem, the
RNG gets seeded and the seed value is reported to make the test
reproducible.