Fix flapping desturctor tests
In perfect conditions, @destructor_order is ["Child", "Parent"] and that was
what the test was looking for. Most of the time however it takes 2 runs of the
loop for the destructors to fire and we end up with: ["Child", "Parent",
"Child", "Parent"]. The test is still fine with that. But sometimes it's
["Child", "Child", "Parent", "Parent"] instead. Which is still ok! Children
destroy before their parents. But the test didn't account for this order.

Fix and improve the test by checking the whole result and making sure we
have correctly ordered pairs of destructor runs. We now also check if we
actually found the correct child for a parent.