Fix error in comment test
The test originally used dies_ok and that passed as the code actually
dies. But it does not do so for the reason it was meant to. Instead of
the invalid comment syntax, it died because of the undeclared variable.
A refactor to move away from dies_ok (for precicely this reason) then
made the wrong test explicit. Fix the test and the expected result to
what it was actually meant to test.