Fix more spurious failures in destructor tests
Destructors may run at any time and concurrently. Protect the
@destructor_order array from concurrent modification using a lock.

Apparently the previous assumption that the WHERE of an object was
stable when we already decided to destroy it was wrong. So instead rely
on default stringification to give a unique and stable identifier (as
Rakudo uses nqp::objectid for constructing the stringification)