Fix accidental re-declaration of our-scoped constant in same package
Constants are our-scoped by default. While an EVAL is a different
lexical scope, it doesn't change the current package. So running two
EVALs, each declaring the same our-scoped constant should yield a
redeclaration error. However, rakudo did not check for this condition
and silently replaced the old stash entry with the new one.