Change tests for column dividers as text in pod tables
To satisfy a more sane behaviour in RakuAST.

In the legacy grammar, there is some magic logic that will interpret
a "+" or "|" at the start of a line with only whitespace to the right
until the next column divider, as a character.  This allows:

=table
+   |  addition
|   |  union
&   |  intersection

to produce:

["+", "addition"], ["|", "union"], ["\&", "intersection"]

Yet, reversing the columns:

=table
set addition     |  +
set union        |  |
set intersection |  &

produces the unexpected:

["addition", "", ""], ["union", "", ""], ["intersection", "\&", ""]

because then the "+" and "|" *are* being considered column
dividers.  And as such produce empty columns.

I've not been able to find documentation on this feature.  After
discussing this with Tom Browder, it felt better to take the rule
for escaping characters that are column dividers (prefix them with
a backslash) the **ONLY** way to have "+" and "|" be interpreted
as characters instead of column dividers.

So when RakuAST has landed, you will have to escape "+" and "|"
inside table cells:

=table
\+  |  addition
\|  |  union
&   |  intersection

to produce:

["+", "addition"], ["|", "union"], ["\&", "intersection"]

It is hoped that this removal of magic will not wreak too much
havoc on pod table rendering.