Mark sprintf tests as v6.e.PREVIEW
But also remove them from the list to check, because with that
change, they are not yet fully passing.  Once they are fully
passing, they will be added back again.

Please note that the original sprintf tests now live in the
6.d directory and are marked as 6.d tests and are still
included.  So the actual coverage has not been changed.