Remove bogus tests for "reference types"
The tests in question, from *2008*, assumed that \ was syntax for
taking a reference to something (like in Perl).  In Raku, the \
syntax is used for creating Captures, which is a different beast
altogether.

The === tests succeeded because Capture was considered a value type
so it had a comparable .WHICH for different Captures with identical
values.

In https://github.com/Raku/problem-solving/issues/372 it was
construed as it **NOT** being a value type.  So these tests will
fail once Captures are no longer value types.

Since these tests were intended to test something that wasn't
implemented in Raku, it feels safe to remove these tests.