Mark failing tests on Sonoma as "todo" for now
These tests appear to fail because of some changes with the "script"
OS utility on Sonoma, possibly related to new security features.
The Raku functionality they try to test, does *not* appear to be
affected, so todoing the tests feels like the right thing to do
just before the 2023.12 release.