Migrate USAGE -> GENERATE-USAGE
GENERATE-USAGE is a part of Raku for a while now and not just
a Rakudo implementation detail.

This patch is a compromise for changing USAGE to conditionally
set the exit status in the same way that GENERATE-USAGE does.

This change causes older USAGE-oriented tests to fail, but rather
than adjust these tests to the new behavior it was deemed a
smoother transition to migrate them to GENERATE-USAGE, which as
stated has been around for quite some time now.

This patch does admittedly re-raise the question of how to
_properly_ account to for version-to-version (or even
release-to-release) shifts in behavior. But rather than revert
a useful and minimal adjustment to USAGE's exit status and blocking
its adoption on remediating this major issue, the approach taken
in this commit was adopted.