Add test for spurt failures
This addresses R#4619. While that ticket focuses on permissions,
I've taken the approach of generating a failure purely based on
trying to spurt a file into a non-existent path.

This feels much less fiddly than what it takes to do a permissions-
specific test (checking for UID 0 and fudging a dummy test if so,
which ignores Windows and also isn't comprehensive).

(https://github.com/rakudo/rakudo/issues/4619)