Make %= test independent of the optimizer
This test would fail when rakudo is run with --optimize=off as the
broken usage of the operator does not throw an exception but instead
stores a Failure. This Failure is stored in the variable. As it then is
inside a container, p6sink won't call .sink on the Failure.

The test worked as intended with the optimizer enabled, as we then
generate a p6assign op which seems to have slightly different semantics.