Fix tests depending on implementation details
Whether the implementation of ~= uses the .Stringy method or the ~
prefix operator to stringify operands is an implementation detail and
indeed even in current Rakudo depends on whether the optimizer is
active, or not. We already check a very specific exception type. There
shouldn't be a need to go beyond that.