Remove bogus test
The old compiler frontend creates an additional lexical scope that's
wrapped around the compunit's scope. Thus OUTER::.keys in an EVAL would
always return an empty list. To access the actual outer scope of the
EVAL, one would have to access OUTER::OUTER. Thus this test did not
actually check whether the variable could be found in the intended
scope.

There's no documentation about what the intent of this test was. The
linked ticket only mentions my $a if 0; crashing the compiler, the
commit message only references this ticket and the test description does
not mention why we shouldn't be able to find the variable.

That the looked-for variable is declared after the EVAL makes me suspect
that the intention was to check whether we'd only find variables defined
before the EVAL lexically. But that seems wrong. OUTER:: was supposed to
reference that lexical scope and the access does not contain any more
position information. Since the scope contains that variable somewhere,
we should find it here. We only didn't because of that bug in the test.

Since the test is buggy and if fixed would actually fail and we don't
even know for sure why it was there or why it the specified behavior was
intended, remove the test.