Fix missing semicolon in test
Commit d7f03fcb90d3cc26ae0c284ac879364834539919 introduce a missing
semicolon error. This however did not surface because the following line
starts with a prefix + which without the semicolon was interpreted as an
infix. The result is that the test sub always returned a true value.

However just adding the semicolon made the test fail. That's because due
to the variable assignment in the supplied &picker sub, we ended up
adding that variable's container to the result set instead of the picked
values. Adding the same container repeatedly left only a single element
in the set which made it fail to pass the test condition.

The variable assignment was taken straight out of the blog post that
code originates from. It has nothing to do with the actual test, so just
remove it.