Fix odd number of arguments when assigning to hash error
These tests are looking for a compilation error. However a different
compiler may run the expected check after the faulty assignment is
already done. The test would then fail due to the unexpected error type.

Fix the unintersting error in the test code so we're really testing what
the test is about.