Fix stress test that used .polymod
https://github.com/rakudo/rakudo/commit/c81d9cfe86499587cde45a6a539b7c636a8364b0
changed the semantics of .polymod for 1.  This test was depending on
that behaviour, but was just using the old behaviour to test a scoping
issue, rather than the specific output of .polymod.

Fixed the test by adding a subroutine that mimicked the old behaviour
and using that instead of the .polymod method call.