Fix tests for nested loops and labeled last
As pointed out by @niner++ the old version assumed that the loops were
not lazy. But as mentioned in the comment the old design docs suggested
that the used syntax (having the loop inside parens) would make the
loop lazy -- at least if the loop is the final statement of a statement
list.

This behavior is currently neither fully specced nor implemented by
Rakudo. The tests weren't meant to specify the non-lazy evaluation.
They only purpose was to make sure that the code doesn't crash and that
the label is taken into account.