Fix bogus >>.?foo test
Since >>. tries to create the same data structure as the invocant
(which in this case is an untyped array), the production of Nil
causes the elements in the new array to be set to the default
value, aka Any.

Instead of doing all sorts of silly things in stringifying and
using "is", use "is-deeply" to check the data stucture actually
returned by >>.?foo.

Found while fixing second part of #5778