Weed out .assuming tests (part 1)
As part of work on #2599 and its new implementation, it became clear
that a lot of these tests are testing a particular implementation's
side effects, rather than the inherent functionality.

The new implementation of .assuming produces different signatures
in a lot of cases, because it can create a much smaller Callable
by using signature features, rather than hiding this away in runtime
arrays and hashes, and then flattening that into a capture to bind
to.

Differences are:
- default values are passed on
Before, a "$a = 42" in a assumee's signature would generate a "$a?".
The new logic generates "$a = 42".

- where constraints are passed on
Before, a "$a where foo" would not pass on the constraint, resulting
in "$a" in the signature (it does produce "$a where foo" now).

- nameless parameters get a generated name
Before, a nameless positional such as "$" would also be generated
as "$".  The new logic assigns a name to it, so that it can be
directly referenced in the call.  So "$" becomes "$param.1".
Note that the param name is not legal in source code, but it works
inside the generated Callable ok.  And because it isn't legal, it
can never interfere with the names of other arguments.

Tests were adapted or removed (if they turn out to be duplicates
of others with nameless arguments).

More tests may be adapted or weeded out in the future.