Remove test that would check implementation detail
The old implementation of .assuming would re-create the constant
value in the signature of the generated Callable.  In the new
implementation, it will pass thru the nameless array that contained
the value.  However, to be able to pass it through, it needs a name
to refer to, which is implementation dependent.

Rather than replace one implementation dependent test with another
it feels like a better idea to remove this test.