Assume Hash.of alwaysi defaults to Mu
Commit 7b7bab4a6c in 2013 assumed Any, as a follow up of
https://github.com/rakudo/rakudo/commit/93e96b44e1b9d9 .

However, since then a lot has happened, and an inconsistency
has crept in:

  dd (my %).of;  # Mu
  dd { }.of;     # Mu
  dd :{ }.of;    # Mu

But:

  dd (my %{Str}).of;  # Any

This is inconsistent.  It shouldn't matter how a Hash is
parameterized or set up, it should always have the same default
"of" type.  And it should be "Mu" to allow for Junctions to
be stored as values in a hash.

  my %h = a => 5 | 6;
  say "matched" if %h<a> == 5;  # matched

This commit changes the test for "Any" to "Mu" and TODOs the
test.  In RakuAST this will be fixed to always have "Mu" as the
default "of".