Fix test for return in parameter defaults and skip it
According to the linked issue, parameter default thunks are supposed to
run inside the routine body they're attached to. Thus a return in such a
thunk would be caught by the return handler of this routine.

The test for this however relied on the return handler of the
surrounding outer sub to catch the return from the inner sub's
signature. This contradicts the stated: "rakudo-moar seems to be under
the impression that the parameter list isn't part of the routine. But
it is; at the point the parameters bind, we're already "running" the
routine, and should be able to return from it."

When one fixes the test to actually match the written description, it
throws an exception. Thus we need to skip it for now.