Remove dubious test that only incidentally passed anyway
https://github.com/Raku/old-issue-tracker/issues/1488 is such a great show case
of ad-hoc language design gone wrong. Someone comes across something unexpected,
there's no discussion and when an incidental change makes the problem disappear
this behavior is enshrined in a spec test.

Arguably the previous unexpected behavior is totally correct.

The block in if True { } is a real block, not just a syntactic grouping. That's
how it's advertised as "strangely consistent" and why you can give it a
signature with if True -> $true { }

@_ is a magical auto-declaring variable that contains the surrounding block's
positional arguments. Since the if statement's block is such a block, that's
what it should be declared on.

m: sub foo { dd @_; if True -> $true { dd @_ } }(1, 2)
rakudo-moar 0730199b3: OUTPUT: «===SORRY!=== Error while compiling
<tmp>␤Placeholder variable '@_' cannot override existing signature␤at
<tmp>:1␤------> sub foo { dd @_; if True <HERE>-> $true { dd @_ } }(1, 2)␤»

After all it does complain about that if block having a signature in this example

But wrap it in another sub:
m: sub bar(*@_) { sub foo { dd @_; if True { dd @_ } }(1, 2) }()
rakudo-moar 0730199b3: OUTPUT: «Too many positionals passed; expected 0
arguments but got 2␤  in sub foo at <tmp> line 1␤  in sub bar at <tmp> line 1␤
in block <unit> at <tmp> line 1␤␤»

In short, @_ *should* implicitly be declared on that if block. That this
can be surprising to users is just a side effect of implicitly declaring
such a variable in the first place.