Reduce test to what actually needs testing
Namely: can we call a method in a subclass of Str, an augmented
version of Str, and a re-augmented version of Str.

This all completely bypasses the fact that augment is a compile
time event, so that all tests could actually be placed *before*
any of the augmentation is done in the code.

All in all, pretty dubious tests