Fix unrelated error from error checking test
The test code contains 2 errors and the test relied on the compiler to
report the "interesting" one first (and only that). A compiler may
however also report the redeclaration first (and arguably that's the
more helpful message even). That redeclaration however is not needed for
testing whether X::Composition::NotComposable.message is accessible so
remove it.