Change compilation error test for: *+42:foo
This should *not* be a compilation error, unless it can be proven
that that will never be an infix:<+>($a,$b,:$foo) or other accepting
infix:<+> candidate at runtime.  Which I don't think you can.

Also, an expression such as:   666+42:foo  *is* allowed, and fails
at runtime if there is no candidate available.  And succeeds if there
is.

In RakuAST the compilation error does not exist, and this causes a
test failure.  I've concluded that in fact the compilation error is
in error in this case, so changed the test to see whether the code
survives compilation.  And TODOed it so it survives in legacy grammar