Fix accidental declaration of our scoped role in another role's signature
The declarations made in a role's signature are meant to go into that role's
body. However it's illegal to make our scoped declarations inside roles
as that scope is generic. Thus it should be wrong to declare an our
scoped role in another role's signature and a more picky compiler might
complain.