Fix test failing for unrelated reasons
The versioning test is not about whether we generate a POPULATE method
(that would be an implementation detail). The expectation of there being
a POPULATE method is just to accomodate that the compiler may generate
one. The way the test was written though would fail if the compiler
didn't.

Change the test to cope with a generated constructor being present but
also not fail if it isn't.