Accommodate tests for sqrt(negative number) in 6.e+
The result of sqrt(-1) changed to the complex number i in 6.e.
Accommodate for this by moving these tests to a separate file
with a "use v6.c" included, and change the tests in the original
file to the value expected in 6.e+, and todoing them now to prevent
spectest breakage until 6.e has become the default.