Accommodate tests for sign(Complex) in 6.e+
The result of sign(i) changed to return a complex number rather
than throwing an exception.  Accommodate for this by moving this
tests to a separate file with a "use v6.c" included, and change
the test in the original file to the value expected in 6.e+, and
todoing them now to prevent spectest breakage until 6.e has become
the default.