Add new test generator for grapheme break tests.
The current test file S15-nfg/GraphemeBreakTest.t file deviates from
other UCD-based test files, in that it parses and runs the tests from
the Unicode data file at test time, rather than generating a test script
ahead of time. It seems to me it would be better to keep things
consistent, so I've designed this new script to generate test files in
the usual style.

A big change here is that the grapheme break test now properly handles
the normalization our NFG strings do during the grapheme breaking
process. The old test was mistakenly written to fail completely on the
first grapheme that got normalized, when in fact Unicode defines the
default grapheme break rules to not be affected by normalization. So
it's actually OK to NFC-normalize the individual expected graphemes and
test them against the graphemes the NFG process spits out.

For the sake of consolidating generated file changes into one commit all
to themselves, the newly pregen'd test files will be added, and the old
test file removed, when all the other Unicode tests have been updated.